feat: Add `ArrowUp` and `ArrowDown`, rename all arrows for consistency
