+++ src/components/Icon/components.tsx
@@ -1,7 +1,11 @@
 export { ReactComponent as Account } from './assets/Account.svg';
 export { ReactComponent as Add } from './assets/Add.svg';
 export { ReactComponent as AddFilled } from './assets/AddFilled.svg';
+export { ReactComponent as ArrowDown } from './assets/ArrowDown.svg';
+export { ReactComponent as ArrowLeft } from './assets/ArrowLeft.svg';
+export { ReactComponent as ArrowRight } from './assets/ArrowRight.svg';
 export { ReactComponent as ArrowTopRight } from './assets/ArrowTopRight.svg';
+export { ReactComponent as ArrowUp } from './assets/ArrowUp.svg';
 export { ReactComponent as Binance } from './assets/Binance.svg';
 export { ReactComponent as BinanceChain } from './assets/BinanceChain.svg';
 export { ReactComponent as CaretDown } from './assets/CaretDown.svg';
@@ -20,7 +24,6 @@ export { ReactComponent as Email } from './assets/Email.svg';
 export { ReactComponent as Eye } from './assets/Eye.svg';
 export { ReactComponent as EyeBlocked } from './assets/EyeBlocked.svg';
 export { ReactComponent as Globe } from './assets/Globe.svg';
-export { ReactComponent as LeftArrow } from './assets/LeftArrow.svg';
 export { ReactComponent as Link } from './assets/Link.svg';
 export { ReactComponent as Lock } from './assets/Lock.svg';
 export { ReactComponent as MastercardColor } from './assets/MastercardColor.svg';
@@ -29,7 +32,6 @@ export { ReactComponent as Page404 } from './assets/Page404.svg';
 export { ReactComponent as Page500 } from './assets/Page500.svg';
 export { ReactComponent as PowerOff } from './assets/PowerOff.svg';
 export { ReactComponent as QRCode } from './assets/QRCode.svg';
-export { ReactComponent as RightArrow } from './assets/RightArrow.svg';
 export { ReactComponent as Settings } from './assets/Settings.svg';
 export { ReactComponent as Shield } from './assets/Shield.svg';
 export { ReactComponent as Success } from './assets/Success.svg';
