+++ src/components/Icon/urls.tsx
@@ -1,7 +1,11 @@
 export { default as Account } from './assets/Account.svg';
 export { default as Add } from './assets/Add.svg';
 export { default as AddFilled } from './assets/AddFilled.svg';
+export { default as ArrowDown } from './assets/ArrowDown.svg';
+export { default as ArrowLeft } from './assets/ArrowLeft.svg';
+export { default as ArrowRight } from './assets/ArrowRight.svg';
 export { default as ArrowTopRight } from './assets/ArrowTopRight.svg';
+export { default as ArrowUp } from './assets/ArrowUp.svg';
 export { default as Binance } from './assets/Binance.svg';
 export { default as BinanceChain } from './assets/BinanceChain.svg';
 export { default as CaretDown } from './assets/CaretDown.svg';
@@ -20,7 +24,6 @@ export { default as Email } from './assets/Email.svg';
 export { default as Eye } from './assets/Eye.svg';
 export { default as EyeBlocked } from './assets/EyeBlocked.svg';
 export { default as Globe } from './assets/Globe.svg';
-export { default as LeftArrow } from './assets/LeftArrow.svg';
 export { default as Link } from './assets/Link.svg';
 export { default as Lock } from './assets/Lock.svg';
 export { default as MastercardColor } from './assets/MastercardColor.svg';
@@ -29,7 +32,6 @@ export { default as Page404 } from './assets/Page404.svg';
 export { default as Page500 } from './assets/Page500.svg';
 export { default as PowerOff } from './assets/PowerOff.svg';
 export { default as QRCode } from './assets/QRCode.svg';
-export { default as RightArrow } from './assets/RightArrow.svg';
 export { default as Settings } from './assets/Settings.svg';
 export { default as Shield } from './assets/Shield.svg';
 export { default as Success } from './assets/Success.svg';
